requirements.txt
gitpython
fs